[TASK] Clean several issues (#3)
Fixes some checks revealed by Scrutinizer.

+++ Classes/Notification/TCA/EntityTcaWriter.php
@@ -251,7 +251,7 @@ abstract class EntityTcaWriter implements SingletonInterface
                     'checkbox' => 0,
                     'default' => 0,
                     'range' => [
-                        'lower' => mktime(0, 0, 0, date('m'), date('d'), date('Y'))
+                        'lower' => mktime(0, 0, 0, (int)date('m'), (int)date('d'), (int)date('Y'))
                     ]
                 ]
             ],
@@ -267,7 +267,7 @@ abstract class EntityTcaWriter implements SingletonInterface
                     'checkbox' => 0,
                     'default' => 0,
                     'range' => [
-                        'lower' => mktime(0, 0, 0, date('m'), date('d'), date('Y'))
+                        'lower' => mktime(0, 0, 0, (int)date('m'), (int)date('d'), (int)date('Y'))
                     ]
                 ]
             ],
